\name{MuStatsData-package}
\alias{MuStatsData-package}
\alias{MuStatsData}
\docType{package}
\title{
\packageTitle{MuStatsData}
}
\description{
\packageDescription{MuStatsData}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MuStatsData}
\packageIndices{MuStatsData}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MuStatsData}

Maintainer: \packageMaintainer{MuStatsData}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
